agent_marketplace/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
cryptography==41.0.7
python-dotenv==1.0.0
PyJWT==2.8.0
asgiref==3.8.1
Brotli==1.1.0
brotlicffi==1.1.0.0
cffi==1.17.1
chardet==5.2.0
colorama==0.4.6
curio==1.6
cython==3.1.2
email_validator==2.2.0
eval_type_backport==0.2.2
exceptiongroup==1.3.0
hypothesis==6.135.26
ipython==8.12.3
outcome==1.3.0.post0
Pygments==2.19.2
pyOpenSSL==25.1.0
pytest==8.3.5
python_bcrypt==0.3.2
python_socks==2.7.1
rich==14.0.0
setuptools==75.8.0
simplejson==3.20.1
socksio==1.0.0
speedups==1.4.0
thread==2.0.5
toml==0.10.2
tomli==2.2.1
uvloop==0.21.0
watchdog==6.0.0
zstandard==0.23.0
requests==2.32.3
supabase==2.16.0